% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockDataGen_Functions.R
\name{generate_addresses}
\alias{generate_addresses}
\title{Generate Fake Addresses}
\usage{
generate_addresses(n = 10)
}
\arguments{
\item{n}{Number of addresses}
}
\value{
A character vector of fake addresses
}
\description{
Builds full fake addresses using manually defined components.
}
