% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockDataGen_Functions.R
\name{generate_ages}
\alias{generate_ages}
\title{Generate Random Ages}
\usage{
generate_ages(n = 10, min_age = 18, max_age = 65)
}
\arguments{
\item{n}{Number of ages}

\item{min_age}{Minimum age (inclusive)}

\item{max_age}{Maximum age (inclusive)}
}
\value{
Integer vector of ages
}
\description{
Returns an integer vector of random ages within a specified range.
}
