% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockDataGen_Functions.R
\name{generate_emails}
\alias{generate_emails}
\title{Generate Fake Email Addresses}
\usage{
generate_emails(
  n = 10,
  domain_pool = c("gmail.com", "yahoo.com", "hotmail.com", "outlook.com", "icloud.com",
    "aol.com", "protonmail.com", "zoho.com", "mail.com", "gmx.com", "yandex.com",
    "live.com")
)
}
\arguments{
\item{n}{Number of email addresses to generate}

\item{domain_pool}{Optional vector of domains to use}
}
\value{
A character vector of email addresses
}
\description{
Uses generated names and appends random email domains.
}
