% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockDataGen_Functions.R
\name{generate_names}
\alias{generate_names}
\title{Generate Fake Names}
\usage{
generate_names(n = 10)
}
\arguments{
\item{n}{Number of names to generate}
}
\value{
A character vector
}
\description{
Returns a character vector of fake names.
}
